# api/docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.RegisterRes:
    properties:
      content:
        type: string
    type: object
...
  models.StandartError:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
...
  models.UserCreate:
    properties:
      card:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      full_name:
        type: string
      gender:
        type: string
      password:
        type: string
      phone_number:
        type: string
      profile_img:
        type: string
    type: object
  models.UserReq:
    properties:
      card:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      password:
        type: string
      phone_number:
        type: string
      profile_img:
        type: string
    type: object
  models.UserRes:
    properties:
      card:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      phone_number:
        type: string
      profile_img:
        type: string
      refresh_token:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  user.ListUsersRes:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/user.User'
        type: array
    type: object
  user.User:
    properties:
      card:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      password:
        type: string
      phone_number:
        type: string
      profile_img:
        type: string
      refresh_token:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: API for Touristan
  title: Welcome To Booking API
paths:
  /v1/admins:
    post:
      consumes:
      - application/json
      description: Api for Create admin
      parameters:
      - description: createModel
        in: body
        name: Admin
        required: true
        schema:
          $ref: '#/definitions/models.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: CREATE ADMIN
      tags:
      - ADMIN
    put:
      consumes:
      - application/json
      description: Api for Update
      parameters:
      - description: createModel
        in: body
        name: Admin
        required: true
        schema:
          $ref: '#/definitions/models.UserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: UPDATE ADMIN
      tags:
      - ADMIN
  /v1/admins/{id}:
    delete:
      consumes:
      - application/json
      description: Api for Delete
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RegisterRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: DELETE ADMIN
      tags:
      - ADMIN
    get:
      consumes:
      - application/json
      description: Api for Get
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: GET ADMIN
      tags:
      - ADMIN
  /v1/admins/list:
    get:
      consumes:
      - application/json
      description: Api for ListAdmins
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ListUsersRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: LIST ADMINS
      tags:
      - ADMIN
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"